requirements.txt
Flask
Flask-Cors
ultralytics
opencv-python
numpy
Pillow